# web/static/openapi.yaml
openapi: 3.0.3
info:
  title: freehire public API
  version: 1.1.0
  description: |
    Public, read-only freehire API: search IT jobs, read the live filter
    vocabulary, inspect job and company details, and find similar jobs. Every
    operation here is unauthenticated. The authenticated tracking surface
    (saved jobs, application stages, CV tooling) is deliberately excluded — this
    schema is the integration contract for search, and is also what the freehire
    custom GPT imports as an Action.

    ## Filter grammar

    Facet parameters are **repeatable**, and the same grammar applies to all of
    them:

    * `?skills=go&skills=kubernetes` — values of one facet are **ORed**: match
      either.
    * `?skills=go&skills=kubernetes&skills_mode=and` — `<facet>_mode=and` flips
      that facet to **AND**: match both. Not available for geography (see below).
    * `?source_exclude=adzuna` — `<facet>_exclude` removes matches. **Every**
      string facet in this schema supports its `_exclude` twin, whether or not it
      is listed below; only the most commonly used ones are declared explicitly.
    * Different facets **AND** together: `?work_mode=remote&seniority=senior` is
      remote *and* senior.

    ## Geography is one OR-group

    `regions`, `countries` and `cities` name a single concept — *where* — so
    their values join into **one OR-group** instead of intersecting.
    `?countries=gb&cities=London` therefore returns *United Kingdom **or**
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
...
paths:
  /jobs/search:
    get:
      operationId: searchJobs
      tags:
        - Jobs
      summary: Search open jobs
      description: |
        Full-text and faceted search over open postings. Descriptions are the
        search index's truncated preview. Use `agentSearchJobs` when you need the
        full verbatim body of every hit, or `getJob` for the one posting the user
        picked.

        Closed postings are never in the index, so they never appear here.
      x-openai-isConsequential: false
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/QFields"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/Regions"
        - $ref: "#/components/parameters/Countries"
        - $ref: "#/components/parameters/Cities"
        - $ref: "#/components/parameters/WorkMode"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Seniority"
        - $ref: "#/components/parameters/Skills"
        - $ref: "#/components/parameters/SkillsMode"
        - $ref: "#/components/parameters/IsTech"
        - $ref: "#/components/parameters/RequiresClearance"
        - $ref: "#/components/parameters/AutoApplyAvailable"
        - $ref: "#/components/parameters/AIArchetype"
        - $ref: "#/components/parameters/RoleType"
        - $ref: "#/components/parameters/Collections"
        - $ref: "#/components/parameters/Reality"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/CompanySlug"
        - $ref: "#/components/parameters/EmploymentType"
        - $ref: "#/components/parameters/Relocation"
        - $ref: "#/components/parameters/EnglishLevel"
        - $ref: "#/components/parameters/EducationLevel"
        - $ref: "#/components/parameters/PostingLanguage"
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CompanyType"
        - $ref: "#/components/parameters/CompanySize"
        - $ref: "#/components/parameters/SalaryCurrency"
        - $ref: "#/components/parameters/SalaryPeriod"
        - $ref: "#/components/parameters/VisaSponsorship"
        - $ref: "#/components/parameters/SalaryMin"
        - $ref: "#/components/parameters/SalaryMax"
        - $ref: "#/components/parameters/ExperienceYearsMin"
        - $ref: "#/components/parameters/ExperienceYearsMax"
        - $ref: "#/components/parameters/PostedWithinDays"
        - $ref: "#/components/parameters/OpenWithinDays"
        - $ref: "#/components/parameters/RegionsExclude"
        - $ref: "#/components/parameters/CountriesExclude"
        - $ref: "#/components/parameters/WorkModeExclude"
        - $ref: "#/components/parameters/RoleTypeExclude"
        - $ref: "#/components/parameters/SkillsExclude"
        - $ref: "#/components/parameters/SourceExclude"
        - $ref: "#/components/parameters/CompanySlugExclude"
      responses:
        "200":
          description: Matching jobs with pagination metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListEnvelope"
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  parameters:
    Q:
      name: q
      in: query
      schema:
        type: string
      description: |
        Full-text query over `title`, `company`, `description`, and `location`.

        Unquoted, multiple words match as an OR of (stemmed) tokens: any word
        matching in any of those fields is a hit. Double-quoted (e.g.
        `q="systems engineer"`), the words match as an AND, irrespective of
        their order or adjacency in the matched field — quoting does **not**
        request a contiguous phrase match. Use `q_fields` to narrow which
        fields `q` is matched against.

        Exact-phrase matching is not offered today, and the default relevance
        order (`sort` omitted) does not reliably rank a contiguous match above
        a scattered-token one within the same field — the index trades that
        precision for cheaper indexing over long-form descriptions. A caller
        that needs phrase precision must currently over-fetch and filter or
        rank client-side.
    QFields:
      name: q_fields
      in: query
      schema:
        type: string
      description: |
        Restrict `q` matching to a comma-separated subset of its four
        searchable fields: `title`, `company`, `description`, `location`. For
        example `q_fields=title` matches "systems" only in the job title, not
        in the company name or body — the only lever `q` has to exclude a
        posting like "STS Systems Defense" from a search for systems engineer
        roles.

        Naming any field outside that set (e.g. `q_fields=salary`) drops the
        parameter entirely — `q` runs unrestricted, across all four fields —
        and `q_fields` is reported in `meta.ignored_params`, the same
        convention every other unrecognized filter follows.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Page size. Values above the maximum are clamped, not rejected.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        # The window is on offset+limit, so this is the ceiling only for the smallest
        # possible limit. A generator cannot express the real bound; the description is
        # what states it.
        maximum: 10000
        default: 0
      description: |
        Rows to skip. On every list endpoint `offset + limit` may not exceed
        10000; deeper paging is a 400 ("pagination too deep"), never a clamped
        page — a 200 carrying rows you did not ask for would leave a client
        walking pages unable to tell it had reached the end. Narrow with
        filters rather than paging that far.
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum:
          - created_at
          - posted_at
          - view_count
          - salary_min
          - salary_max
      description: |
        Sort field. Omitted, a text query sorts by relevance and an empty query
        by `posted_at` descending. `view_count` orders by how many people have
        opened each posting.
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
      description: Sort direction. Ignored without a valid `sort`.
...
    Regions:
      name: regions
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - global
            - north_america
            - latam
            - eu
            - uk
            - mena
            - africa
            - apac
            - cis
            - none
      description: |
        Macro-regions. Country codes are NOT regions — use `countries`. `none`
        is the reserved value for postings with no resolved geography. Joins the
        geography OR-group with `countries` and `cities`.
    Countries:
      name: countries
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        ISO 3166-1 alpha-2, lowercase (`gb`, `de`). Matching is
        case-insensitive. Joins the geography OR-group.
    Cities:
      name: cities
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Canonical city display names ("London", "Berlin") — resolve one with
        `searchCities` first. Joins the geography OR-group.
    WorkMode:
      name: work_mode
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - remote
            - hybrid
            - onsite
      description: |
        Work format, resolved from a deterministic dictionary. Absent on a
        posting whose format could not be resolved, so this filter narrows to
        postings that stated it.
    Category:
      name: category
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Role category slug from `getJobFacets`.
    Seniority:
      name: seniority
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - intern
            - junior
            - middle
            - senior
            - lead
            - staff
            - principal
            - c_level
    Skills:
      name: skills
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Canonical skill slugs from `getJobFacets`. Never invent one.
    SkillsMode:
      name: skills_mode
      in: query
      schema:
        type: string
        enum:
          - and
      description: |
        `and` requires a posting to carry EVERY listed skill instead of any of
        them. The same `<facet>_mode=and` switch works on any non-geography
        facet.
    RequiresClearance:
      name: requires_clearance
      in: query
      schema:
        type: boolean
      description: |
        Whether the posting states a government security-clearance requirement
        (UK SC/DV/BPSS, US Secret/TS-SCI/polygraph, AU baseline/NV1), derived
        deterministically from the description.

        `false` returns every posting NOT marked as requiring one — including the
        vast majority, where the description says nothing about clearance. It is
        the "hide the jobs I cannot be hired into" filter, not an assertion that
        those postings promise no clearance is needed.

        `true` returns only the marked postings, for a cleared candidate seeking
        exactly that work. Omitting the parameter filters nothing.
    AutoApplyAvailable:
      name: auto_apply_available
      in: query
      schema:
        type: boolean
      description: |
        Whether the posting's ATS provider is one freehire's auto-apply feature
        can currently attempt to fill and submit for. A best-effort,
        provider-eligibility signal only — a marked posting can still park on a
        live obstacle (a captcha, an unrecognized form layout, missing
        candidate answers) discovered only when an attempt actually runs, never
        a guarantee of successful submission.

        `true` returns only the marked postings. Omitting the parameter filters
        nothing.
    IsTech:
      name: is_tech
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - tech
            - non_tech
      description: |
        Technical vs non-technical, derived deterministically from title and
        category. Absent when unknown, so the filter never guesses.
    AIArchetype:
      name: ai_archetype
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - rag_app_builder
            - agent_builder
            - cloud_ml_platform_engineer
            - ml_trainer_researcher
            - fullstack_ai_engineer
            - devops_infra_engineer
      description: AI skill-signature archetype, derived from the posting's skill set.
    RoleType:
      name: role_type
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - people_manager
      description: |
        Whether the posting names a people-management role, resolved from its title.

        The vocabulary holds ONE value on purpose. A posting carrying
        `people_manager` is a management role; a posting carrying **nothing** means
        no management marker was found in the title — it does **not** mean the
        posting is an individual-contributor role. That side is not detectable from
        a title and this API does not claim it.

        So `role_type_exclude=people_manager` returns "postings with no management
        marker", which is a wider and weaker set than "individual-contributor
        postings". Do not present it to users as the latter.
    Collections:
      name: collections
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Curated collection slugs of the posting's company, e.g. `yc`,
        `bigtech`, `unicorn`, `us-h1b-sponsor`. Read the live set from
        `getJobFacets`.
    Reality:
      name: reality
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - fresh
            - stale
            - likely-evergreen
      description: |
        Posting-reality class. `fresh` is recently posted and not obviously
        recycled; `stale` is old; `likely-evergreen` reads as an always-open
        pipeline ad. A large share of the catalogue is `stale`, so `?reality=fresh`
        is the cheapest quality filter available. Read the live split from
        `getJobFacets`.
    Source:
      name: source
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Origin slug of the posting — the ATS or board it was crawled from
        (`greenhouse`, `workday`, `adzuna`, …). Read the live set from
        `getJobFacets`. Pair with `source_exclude` to avoid double-counting a
        board you already query directly.
    CompanySlug:
      name: company_slug
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: |
        Company slugs from `searchCompanies`. This facet has no distribution in
        `getJobFacets` — use the company search for a typeahead.
    EmploymentType:
      name: employment_type
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - full_time
            - part_time
            - contract
            - internship
            - fellowship
    Relocation:
      name: relocation
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - not_supported
            - supported
            - required
    EnglishLevel:
      name: english_level
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - none
            - a1
            - a2
            - b1
            - b2
            - c1
            - c2
            - native
    EducationLevel:
      name: education_level
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - none
            - bachelor
            - master
            - phd
    PostingLanguage:
      name: posting_language
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Language the posting is written in (ISO 639-1, e.g. `en`, `de`, `uk`).
    Domains:
      name: domains
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - fintech
            - crypto
            - ecommerce
            - gambling
            - gamedev
            - media
            - travel
            - healthcare
            - edtech
            - govtech
            - devtools
            - cybersecurity
            - ai
            - hrtech
            - adtech
            - proptech
            - logistics
            - mobility
            - climatetech
            - other
      description: Business domain of the hiring company, as stated by the posting.
    CompanyType:
      name: company_type
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - product
            - startup
            - outsource
            - outstaff
            - agency
            - inhouse
            - government
    CompanySize:
      name: company_size
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - 1-10
            - 11-50
            - 51-200
            - 201-500
            - 501-1000
            - 1000+
    SalaryCurrency:
      name: salary_currency
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: ISO 4217 code, e.g. `USD`, `EUR`.
    SalaryPeriod:
      name: salary_period
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - year
            - month
            - day
            - hour
    VisaSponsorship:
      name: visa_sponsorship
      in: query
      schema:
        type: boolean
      description: |
        Whether the posting states it sponsors a visa. `false` is a real,
        stated value, not "unknown" — postings that say nothing match neither
        setting.
    SalaryMin:
      name: salary_min
      in: query
      schema:
        type: integer
      description: |
        Lower bound on the posting's minimum salary, in its own
        `salary_currency` and `salary_period` — the comparison does NOT
        normalize currencies. Pair with `salary_currency` and `salary_period` for
        a meaningful range.
    SalaryMax:
      name: salary_max
      in: query
      schema:
        type: integer
      description: Upper bound on the posting's maximum salary. Same currency caveat as `salary_min`.
    ExperienceYearsMin:
      name: experience_years_min
      in: query
      schema:
        type: integer
      description: >-
        Lower bound on the years of experience the posting asks for. Pairs with
        `experience_years_max` as a closed range over that one figure. A posting
        that states no experience requirement carries no figure and is therefore
        excluded by either bound.
    ExperienceYearsMax:
      name: experience_years_max
      in: query
      schema:
        type: integer
        minimum: 0
      description: >-
        Upper bound on the years of experience the posting asks for — the most you
        are willing to be asked for. It bounds the same figure `experience_years_min`
        bounds from below, not a second field. `experience_years_max=0` selects the
        postings that state no prior experience is required. Like `experience_years_min`,
        it excludes postings that state no requirement at all.
    PostedWithinDays:
      name: posted_within_days
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 36500
      description: >-
        Restrict to postings whose SOURCE states a publication date within the last N
        days. Omit for no restriction. Some boards restate this date on every crawl, so
        a posting that has been open for months can satisfy a narrow bound here — see
        `open_within_days` for the date that cannot be rewritten from outside. The two
        are independent and AND together.
    OpenWithinDays:
      name: open_within_days
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 36500
      description: >-
        Restrict to postings first recorded by this system within the last N days — how
        long the posting has been in the catalogue, regardless of what date its source
        states. Omit for no restriction. This is the bound to use for "nothing that has
        been sitting here for months"; `posted_within_days` answers the different
        question of what the board claims.
    RegionsExclude:
      name: regions_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Regions to exclude. Excludes AND together, unlike included geography.
    CountriesExclude:
      name: countries_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Countries to exclude.
    WorkModeExclude:
      name: work_mode_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Work formats to exclude, e.g. `onsite`.
    RoleTypeExclude:
      name: role_type_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          enum:
            - people_manager
      description: >-
        Role types to exclude. Declared explicitly, unlike most `_exclude` twins,
        because with a one-value vocabulary the exclusion is the primary way to use
        this facet: `role_type_exclude=people_manager` returns the postings whose
        title carries no management marker. That is a wider and weaker set than
        "individual-contributor postings" — it also holds every management role the
        title dictionary missed — so do not present it to users as the latter.
    SkillsExclude:
      name: skills_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Skill slugs to exclude.
    SourceExclude:
      name: source_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Source slugs to exclude.
    CompanySlugExclude:
      name: company_slug_exclude
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Company slugs to exclude.
  schemas:
    Job:
      type: object
      additionalProperties: true
      required:
        - public_slug
        - title
        - company
      properties:
        public_slug:
          type: string
          description: Stable public identifier. The freehire page is https://freehire.me/jobs/{public_slug}.
        title:
          type: string
        company:
          type: string
        company_slug:
          type: string
        url:
          type: string
          format: uri
          description: The employer's own application URL. This is where an applicant goes.
        location:
          type: string
          description: The posting's own location string, verbatim and unnormalized. Filter on the geography facets instead.
        source:
          type: string
          description: Origin slug — the ATS or board this posting was crawled from.
        external_id:
          type: string
          description: The source's own identifier for the posting.
        manually_added:
          type: boolean
          description: True for a hand-curated posting rather than an automated crawl.
        description:
          type: string
          description: |
            Stored verbatim HTML by default. `searchJobs` serves a truncated
            preview; `agentSearchJobs` and `getJob` serve the full body, and
            `agentSearchJobs` can convert it via `description_format`.
        regions:
          type: array
          items:
            type: string
        countries:
          type: array
          items:
            type: string
        cities:
          type: array
          items:
            type: string
        work_mode:
          type: string
          enum:
            - remote
            - hybrid
            - onsite
          description: Absent when the posting's format could not be resolved.
        skills:
          type: array
          items:
            type: string
        collections:
          type: array
          items:
            type: string
          description: Curated collection slugs of the posting's company.
        is_tech:
          type: string
          enum:
            - tech
            - non_tech
          description: Absent when unknown.
        requires_clearance:
          type: boolean
          description: |
            Present and `true` only when the posting states a government
            security-clearance requirement. Absent otherwise — there is no
            `false`, because "no requirement was detected" is not the same claim
            as "this posting needs no clearance".
        auto_apply_available:
          type: boolean
          description: |
            Present and `true` only when the posting's ATS provider is one
            freehire's auto-apply feature can currently attempt to fill and
            submit for. Absent otherwise — a best-effort provider-eligibility
            signal, never a guarantee of successful submission.
        posted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
          description: When freehire first saw the posting.
        updated_at:
          type: string
          format: date-time
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
          description: When a re-crawl last confirmed the posting was still live.
        closed_at:
          type: string
          format: date-time
          nullable: true
          description: Non-null once the posting is no longer open. Only `getJob` ever serves such a row.
        enrichment:
          $ref: "#/components/schemas/Enrichment"
        enriched_at:
          type: string
          format: date-time
          nullable: true
        enrichment_version:
          type: integer
          nullable: true
          description: Bumped when the enrichment contract changes — use it to detect re-enriched rows.
        my_vote:
          type: integer
          nullable: true
          description: The calling user's vote. Always null on this unauthenticated schema.
        reality:
          $ref: "#/components/schemas/Reality"
        ghost:
          $ref: "#/components/schemas/Ghost"
        view_count:
          type: integer
        applied_count:
          type: integer
        upvote_count:
          type: integer
        downvote_count:
          type: integer
    Enrichment:
      type: object
      additionalProperties: true
      description: |
        Model-extracted attributes, every field omitted when the posting did not
        state it. `work_mode`, `regions`, `countries`, `cities` and `skills` are
        deliberately NOT here — those are served top-level from the
        deterministic dictionaries, and the model's own values for them are not
        published.
      properties:
        summary:
          type: string
          description: The one synthesized field — a 1-2 sentence synopsis, always present on an enriched posting.
        employment_type:
          type: string
        relocation:
          type: string
        visa_sponsorship:
          type: boolean
        timezone_note:
          type: string
        salary_min:
          type: integer
        salary_max:
          type: integer
        salary_currency:
          type: string
        salary_period:
          type: string
        seniority:
          type: string
        experience_years_min:
          type: integer
        english_level:
          type: string
        education_level:
          type: string
        category:
          type: string
        domains:
          type: array
          items:
            type: string
        posting_language:
          type: string
        company_type:
          type: string
        company_size:
          type: string
...
    Reality:
      type: object
      additionalProperties: true
      description: Signals about whether the posting reflects a real, current opening.
      properties:
        class:
          type: string
          enum:
            - fresh
            - stale
            - likely-evergreen
        age_days:
          type: integer
        repost_count:
          type: integer
          description: How many times this role has been reposted.
        mass_posting_count:
          type: integer
          description: How many near-identical postings the same role cluster holds.
        fake_freshness:
          type: boolean
          description: True when the stated posting date looks refreshed rather than real.
    JobListEnvelope:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    PaginationMeta:
      type: object
      additionalProperties: true
      properties:
        total:
          type: integer
          description: |
            Postings matching the filter. Exact on the job endpoints; on an
            unfiltered `searchCompanies` request it is a planner estimate.
        limit:
          type: integer
          description: The page size actually applied, after clamping.
        offset:
          type: integer
        ignored_params:
          type: array
          description: |
            Present ONLY when the request carried a parameter no filter read —
            a clean request omits the key entirely rather than sending `[]`.
            **Check it.** A misspelled filter does not fail; it is dropped, and
            the unfiltered result that comes back is indistinguishable from a
            genuine one.
          items:
            $ref: "#/components/schemas/IgnoredParam"
    IgnoredParam:
      type: object
      required:
        - param
      properties:
        param:
          type: string
          description: The query parameter that was received and not used.
        did_you_mean:
          type: string
          description: |
            The vocabulary's name for it, when the only mistake was grammatical
            number — most facets are plural, so `country` resolves to
            `countries`. Absent when nothing is close enough; a guess would
            mislead more than silence.
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string